Cleaned up the code a little and added some styling.
